+++ packages/mongo-nav/src/types.ts
@@ -6,6 +6,7 @@ const NavElement = {
   OrgNavSupport: 'orgNavSupport',
   OrgNavBilling: 'orgNavBilling',
   OrgNavAdmin: 'orgNavAdmin',
+  OrgNavDocs: 'orgNavDocs',
   OrgNavAllClusters: 'orgNavAllClusters',
   Logout: 'logout',
   OrgNavLeaf: 'leaf',
@@ -14,7 +15,7 @@ const NavElement = {
   OrgNavViewAllOrganizations: 'orgNavViewAllOrganizations',
   OrgNavDropdownOrgAccessManager: 'orgNavDropdownOrgAccessManager',
   OrgNavDropdownProjectAccessManager: 'orgNavDropdownProjectAccessManager',
-  OrgNavDropdownMoreMenu: 'orgNavDropdownMoreMenu',
+  OrgNavDropdownGetHelp: 'orgNavDropdownGetHelp',
   ProjectNavProjectSelectTrigger: 'projectNavProjectSelectTrigger',
   ProjectNavProjectSelectSearch: 'projectNavProjectSelectSearch',
   ProjectNavViewAllProjects: 'projectNavViewAllProjects',
@@ -273,6 +274,7 @@ interface OrgNavURLS {
   settings?: string;
   accessManager?: string;
   support?: string;
+  docs?: string;
   billing?: string;
   allClusters?: string;
   admin?: string;