+++ packages/tabs/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @leafygreen-ui/tabs
 
+## 5.1.0
+
+### Minor Changes
+
+- b9a8df1e: Export Props used by Tabs
+
+### Patch Changes
+
+- Updated dependencies [65032024]
+  - @leafygreen-ui/palette@3.2.0
+
+## 5.0.3
+
+### Patch Changes
+
+- b3508a2e: Addresses bug where multiple sets of Tabs on a page, controlled by the same state, would automatically move focus to the last set.
+
 ## 5.0.2
 
 ### Patch Changes